# arenaskl: avoid repeated comparison with known tower